ripngd: implement optional heuristic suggested by RFC 2080
RFC 2080 - Section 2.4.2:
"If the new metric is the same as the old one, examine the timeout for the
existing route.  If it is at least halfway to the expiration point, switch
to the new route.  This heuristic is optional, but highly recommended".

Implement this optional heuristic only when ECMP is disabled globally ("no
allow-ecmp"), otherwise all routes with the same metric should be used.

Fixes IxANVL RIPng test 7.21.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>