bgpd: Allow views to 'pretend' resolve nexthops
Views are supposed to be independent tables
that have no connection to routing tables.
Since View's are 'independent' there is
no way to do 'real' nexthop resolution since
connected routes and interfaces are associated
with a particular table (or really vrf).
So when we have a bgp instance assume that
nexthops specified are actually valid,
since we are propagating what our neighbors are telling
us.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>