eigrpd: Fix communication w/ Zebra
We were not properly sending data to zebra for
route installation.

I removed the metric installation because I am not sure
if we've actually been passed the metric to install.
if we need it it can be added back in later.

Additionally cleaned up some silliness in eigrp_topology.c

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>